requirements.txt
comet_ml==3.49.10
datasets==3.6.0
huggingface_hub==0.31.2
python-dotenv==1.1.0
torch==2.7.0+cu128
transformers==4.51.3
trl==0.17.0
unsloth==2025.5.3